linebotandaichat/requirements.txt
flask
line-bot-sdk==3.*
requests
python-dotenv